Clean up the tests too
